pyproject.toml
[tool.poetry]
name = "dadit"
version = "0.1.0"
description = ""
authors = ["Bob van der Linden <bobvanderlinden@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tree-sitter-languages = "^1.10.2"
ruamel-yaml = "^0.18.6"
jsonpatch = "^1.33"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.356"
pytest = "^8.1.1"

[tool.poetry.scripts]
dadit = "dadit:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
